mini reverse

+++ YukkiMusic/plugins/bot/help.py
@@ -12,7 +12,7 @@ from typing import Union
 
 from pyrogram import filters, types
 from pyrogram.types import InlineKeyboardMarkup, Message
-from config import SUPPORT_GROUP
+
 from config import BANNED_USERS
 from strings import get_command, get_string, helpers
 from YukkiMusic import app
@@ -27,9 +27,7 @@ HELP_COMMAND = get_command("HELP_COMMAND")
 
 
 @app.on_message(filters.command(HELP_COMMAND) & filters.private & ~BANNED_USERS)
-@app.on_callback_query(
-    filters.regex("settings_back_helper") & ~BANNED_USERS
-)
+@app.on_callback_query(filters.regex("settings_back_helper") & ~BANNED_USERS)
 async def helper_private(
     client: app, update: Union[types.Message, types.CallbackQuery]
 ):
@@ -43,22 +41,22 @@ async def helper_private(
         language = await get_lang(chat_id)
         _ = get_string(language)
         keyboard = help_pannel(_, True)
-        await update.edit_message_text(
-            _["help_1"].format(SUPPORT_GROUP), reply_markup=keyboard
-        )
+        if update.message.photo:
+            await update.message.delete()
+            await update.message.reply_text(_["help_1"], reply_markup=keyboard)
+        else:
+            await update.edit_message_text(_["help_1"], reply_markup=keyboard)
     else:
-        try:
-            await update.delete()
-        except:
-            pass
-        language = await get_lang(update.chat.id)
+        chat_id = update.chat.id
+        if await is_commanddelete_on(update.chat.id):
+            try:
+                await update.delete()
+            except:
+                pass
+        language = await get_lang(chat_id)
         _ = get_string(language)
         keyboard = help_pannel(_)
-        await update.reply_photo(
-            photo=START_IMG_URL,
-            caption=_["help_1"].format(SUPPORT_GROUP),
-            reply_markup=keyboard,
-      )
+        await update.reply_text(_["help_1"], reply_markup=keyboard)
 
 
 @app.on_message(filters.command(HELP_COMMAND) & filters.group & ~BANNED_USERS)
